fleet/apps/backend/requirements.txt
Django==5.2.7
djangorestframework==3.16.1
asgiref==3.10.0
sqlparse==0.5.3
